refactor(api): move SubmitQueryRequest to the crate that owns the v1 wire types
`SubmitQueryRequest` is the request body of `POST /v1/queries` — a versioned
public API contract — but it was defined inside `runtime`'s private `http::v1`
module. Two things had to work around that. `jobs` reached back up into
`http::v1::queries` for the type it is handed, and `runtime` carried a
`pub mod http_types` whose only content was a re-export of this one type, so the
cluster integration tests could name it.

Move it to `runtime-api-types`, which exists for exactly this: the canonical
`/v1` request and response types shared by the runtime's handlers and the CLI,
with the stability guarantee those types need. The struct is unchanged — same
fields, serde attributes and derives — and the crate already carries the
`openapi` feature that gates the `utoipa::ToSchema` derive, which `runtime`
already forwards, so no feature wiring changes.

`runtime::http_types` is deleted, which narrows runtime's public API, and `jobs`
no longer depends on `http` at all.
