+++ crates/runtime/src/lib.rs
@@ -106,10 +106,6 @@ pub use runtime_table::federated;
 pub mod flight;
 mod http;
 
-pub mod http_types {
-    pub use crate::http::v1::queries::SubmitQueryRequest;
-}
-
 mod init;
 pub mod internal_table;
 pub mod jobs;
