+++ crates/runtime/tests/cluster/job_store.rs
@@ -24,9 +24,9 @@ use object_store::{ObjectStore, path::Path};
 use runtime::Runtime;
 use runtime::cluster::ResolvedClusterConfig;
 use runtime::config::ClusterConfig;
-use runtime::http_types::SubmitQueryRequest;
 use runtime::jobs::{JobExecutor, JobStore};
 use runtime::{auth::EndpointAuth, config::Config};
+use runtime_api_types::v1::queries::SubmitQueryRequest;
 use rustls::crypto::{CryptoProvider, aws_lc_rs};
 use spicepod::component::dataset::Dataset;
 use std::net::{Ipv4Addr, SocketAddrV4};
