+++ crates/runtime/tests/cluster/scheduler_failover.rs
@@ -57,9 +57,9 @@ use ballista_scheduler::state::executor_manager::ExecutorManager;
 use runtime::auth::EndpointAuth;
 use runtime::cluster::ResolvedClusterConfig;
 use runtime::config::{ClusterConfig, ClusterRole, Config};
-use runtime::http_types::SubmitQueryRequest;
 use runtime::jobs::{JobExecutor, JobState, JobStatus};
 use runtime::{Result as RuntimeResult, Runtime};
+use runtime_api_types::v1::queries::SubmitQueryRequest;
 use rustls::crypto::{CryptoProvider, aws_lc_rs};
 use spicepod::component::dataset::Dataset;
 use spicepod::component::runtime::{Runtime as SpicepodRuntime, Scheduler as SchedulerConfig};
